+++ routes/permission.test.js
@@ -4,9 +4,11 @@ import { describe, it, before, after } from 'node:test'
 import { init } from './index.js'
 import Group from '../lib/group.js'
 import User from '../lib/user.js'
+import Permission from '../lib/permission.js'
 
 import groupCase from './test/group.json' with { type: 'json' }
 import userCase from './test/user.json' with { type: 'json' }
+import permCase from './test/permission.json' with { type: 'json' }
 
 let server
 
@@ -14,13 +16,14 @@ before(async () => {
   server = await init()
   await Group.create(groupCase)
   await User.create(userCase)
+  await Permission.create(permCase)
 })
 
 after(async () => {
   await server.stop()
 })
 
-describe('group routes', () => {
+describe('permission routes', () => {
   let sessionCookie
 
   it('POST /session establishes a session', async () => {
@@ -36,30 +39,34 @@ describe('group routes', () => {
     sessionCookie = res.headers['set-cookie'][0].split(';')[0]
   })
 
-  it(`GET /group/${groupCase.id}`, async () => {
+  it(`GET /permission/${userCase.id}`, async () => {
     const res = await server.inject({
       method: 'GET',
-      url: `/group/${groupCase.id}`,
+      url: `/permission/${userCase.id}`,
       headers: {
         Cookie: sessionCookie,
       },
     })
     // console.log(res.result)
     assert.equal(res.statusCode, 200)
+    assert.equal(res.result.permission.zone.create, true)
+    assert.equal(res.result.permission.nameserver.create, false)
   })
 
-  const case2Id = 4094
+  let case2Id = 4094
 
-  it('POST /group', async () => {
-    const testCase = JSON.parse(JSON.stringify(groupCase))
+  it('POST /permission', async () => {
+    const testCase = JSON.parse(JSON.stringify(permCase))
     testCase.id = case2Id // make it unique
-    testCase.name = `example2.com`
+    testCase.uid = case2Id
+    testCase.gid = case2Id
+    testCase.name = `Route Test Permission 2`
     delete testCase.deleted
     // console.log(testCase)
 
     const res = await server.inject({
       method: 'POST',
-      url: '/group',
+      url: '/permission',
       headers: {
         Cookie: sessionCookie,
       },
@@ -67,24 +74,28 @@ describe('group routes', () => {
     })
     // console.log(res.result)
     assert.equal(res.statusCode, 201)
+    assert.equal(res.result.permission.zone.create, true)
+    assert.equal(res.result.permission.nameserver.create, false)
   })
 
-  it(`GET /group/${case2Id}`, async () => {
+  it(`GET /permission/${case2Id}`, async () => {
     const res = await server.inject({
       method: 'GET',
-      url: `/group/${case2Id}`,
+      url: `/permission/${case2Id}`,
       headers: {
         Cookie: sessionCookie,
       },
     })
     // console.log(res.result)
     assert.equal(res.statusCode, 200)
+    assert.equal(res.result.permission.zone.create, true)
+    assert.equal(res.result.permission.nameserver.create, false)
   })
 
-  it(`DELETE /group/${case2Id}`, async () => {
+  it(`DELETE /permission/${case2Id}`, async () => {
     const res = await server.inject({
       method: 'DELETE',
-      url: `/group/${case2Id}`,
+      url: `/permission/${case2Id}`,
       headers: {
         Cookie: sessionCookie,
       },
@@ -93,22 +104,23 @@ describe('group routes', () => {
     assert.equal(res.statusCode, 200)
   })
 
-  it(`GET /group/${case2Id}`, async () => {
+  it(`GET /permission/${case2Id}`, async () => {
     const res = await server.inject({
       method: 'GET',
-      url: `/group/${case2Id}`,
+      url: `/permission/${case2Id}`,
       headers: {
         Cookie: sessionCookie,
       },
     })
     // console.log(res.result)
-    assert.equal(res.statusCode, 204)
+    // assert.equal(res.statusCode, 200)
+    assert.equal(res.result.permission, undefined)
   })
 
-  it(`GET /group/${case2Id} (deleted)`, async () => {
+  it(`GET /permission/${case2Id} (deleted)`, async () => {
     const res = await server.inject({
       method: 'GET',
-      url: `/group/${case2Id}?deleted=1`,
+      url: `/permission/${case2Id}?deleted=1`,
       headers: {
         Cookie: sessionCookie,
       },
@@ -117,10 +129,10 @@ describe('group routes', () => {
     assert.equal(res.statusCode, 200)
   })
 
-  it(`DELETE /group/${case2Id}`, async () => {
+  it(`DELETE /permission/${case2Id}`, async () => {
     const res = await server.inject({
       method: 'DELETE',
-      url: `/group/${case2Id}?destroy=true`,
+      url: `/permission/${case2Id}?deleted=1&destroy=true`,
       headers: {
         Cookie: sessionCookie,
       },