+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### 3.0.0-alpha.3
 
-- 
+- permission.get: default search with deleted=0
+
 
 
 [3.0.0-alpha.3]: https://github.com/NicTool/api/releases/tag/3.0.0-alpha.3